# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    CodebaseSourceType:
      type: string
      enum: [local, github]
...
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          const: true
...
paths:
...
  /api/clone/local:
    post:
      operationId: loadLocalRepo
      summary: Load an existing local folder (git optional)
      description: >-
        Validates and loads any existing, readable local directory as a
        project. Being a git repository is optional: `git` reports whether
        the folder is a git working repository, and branch/status metadata
        is only populated when it is.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: Loaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  name:
                    type: string
                  path:
                    type: string
                  git:
                    type: boolean
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
...
  /api/workspaces/{id}/codebases:
    get:
      operationId: listWorkspaceCodebases
      summary: List codebases in a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Codebase list
          content:
            application/json:
              schema:
                type: object
                properties:
                  codebases:
                    type: array
                    items:
                      type: object
    post:
      operationId: addWorkspaceCodebase
      summary: Add a codebase to a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                path:
                  type: string
                sourceType:
                  $ref: "#/components/schemas/CodebaseSourceType"
      responses:
        "201":
          description: Created codebase
...
  /api/codebases/{id}:
    patch:
      operationId: updateCodebase
      summary: Update codebase metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
    delete:
      operationId: deleteCodebase
      summary: Delete a codebase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...